refactor : separate out test cases for filter and util
